Moved ClassTest to correct folder
