+++ tsconfig.json
@@ -98,7 +98,7 @@
 		"source/Tokens/Tests/SeparatorTest.ts",
 		"source/Tokens/Tests/WhitespaceTest.ts",
 
-		"source/SyntaxTree/Tests/ClassTest.ts",
+		"source/SyntaxTree/Declarations/Tests/ClassTest.ts",
 
 		"source/Utilities/Tests/ArrayIteratorTest.ts",
 		"source/Utilities/Tests/BufferedIteratorTest.ts",
